#251: add/remove docent/tutor in course creation and course editing

+++ web-gui/src/app/components/courses/grant-docent-snipp/grant-docent-snipp.component.ts
@@ -15,6 +15,8 @@ import {Observable} from "rxjs";
 export class GrantDocentSnippComponent implements OnInit {
   @ViewChild(MatSort) sort: MatSort;
   @Input() course;
+  @Input() docent_list: User[];
+  @Input() liveUpdate:boolean;
   @Output() loadAllCourses: EventEmitter<void>;
 
   docentFormControl = new FormControl();
@@ -65,10 +67,22 @@ export class GrantDocentSnippComponent implements OnInit {
       const selectedUser: User = this.docentFormControl.value;
       this.docentFormControl.setValue('');
       this.showInputForDocent = false;
+      if(this.liveUpdate) {
+        this.db.addDocentToCourse(courseID, selectedUser.user_id).subscribe(res => {
+          this.loadAllCourses.emit();
+        })
+      } else {
+        this.docent_list.push(selectedUser)
+        console.log(this.docent_list)
+      }
+    }
+  }
 
-      this.db.addDocentToCourse(courseID, selectedUser.user_id).subscribe(res => {
-        this.loadAllCourses.emit();
-      })
+  get correctCourseDocent(){
+    if(this.liveUpdate) {
+      return this.course.course_docent
+    } else {
+      return this.docent_list
     }
   }
 
@@ -78,9 +92,20 @@ export class GrantDocentSnippComponent implements OnInit {
    * @param userID The docent id
    */
   removeDocent(courseID: number, userID: number) {
-    this.db.removeDocentFromCourse(courseID, userID).subscribe(courses => {
-      this.loadAllCourses.emit();
-    });
+    if(this.liveUpdate){
+      this.db.removeDocentFromCourse(courseID, userID).subscribe(courses => {
+        this.loadAllCourses.emit();
+      });
+    } else {
+      let hiddenUser = this.docent_list.filter((u: User) => {
+        return u.user_id == userID
+      })
+      console.log("here",hiddenUser )
+      let i = this.docent_list.indexOf(hiddenUser[0])
+      this.docent_list.splice(i,1)
+
+    }
+
   }
 
   /**
